requirements.txt
Flask==3.0.0
flask_sqlalchemy==3.1.1
Requests==2.31.0
psycopg2-binary==2.9.9
